refactor: reorganize the models in app form
Divide the models into your own model sets that the management of the
models is better and the sets are easier to understand.
