+++ application_form/models/__init__.py
@@ -1,0 +1,22 @@
+from application_form.models.application import (
+    Applicant,
+    Application,
+    ApplicationApartment,
+)
+from application_form.models.lottery import LotteryEvent, LotteryEventResult
+from application_form.models.reservation import (
+    ApartmentQueue,
+    ApartmentQueueApplication,
+    ApartmentQueueChangeEvent,
+)
+
+__all__ = [
+    "Applicant",
+    "Application",
+    "ApplicationApartment",
+    "LotteryEvent",
+    "LotteryEventResult",
+    "ApartmentQueue",
+    "ApartmentQueueApplication",
+    "ApartmentQueueChangeEvent",
+]
