+++ application_form/tests/test_lottery_hitas_services.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from apartment.tests.factories import ApartmentFactory, ProjectFactory
 from application_form.enums import ApplicationState, ApplicationType
-from application_form.models import LotteryEventResult
+from application_form.models.lottery import LotteryEventResult
 from application_form.services.application import (
     cancel_hitas_application,
     get_ordered_applications,
