+++ application_form/tests/test_queue_services.py
@@ -4,7 +4,10 @@ from unittest.mock import Mock
 
 from apartment.tests.factories import ApartmentFactory
 from application_form.enums import ApartmentQueueChangeEventType, ApplicationType
-from application_form.models import ApartmentQueue, ApartmentQueueChangeEvent
+from application_form.models.reservation import (
+    ApartmentQueue,
+    ApartmentQueueChangeEvent,
+)
 from application_form.services.application import get_ordered_applications
 from application_form.services.queue import (
     add_application_to_queues,
